Fallback on OANDA data set. Better error msgs
Use Faraday to capture and return more indicative error responses.
Catch the error that may be caused by a missing currency in our first
requested dataset, try another request with OANDA data set and only
raise exception if even that fails. Previously, only open-uri#read was
used, which raises a generic 400 Bad Request that is not very helpful
for debugging. This prevents code to go to the next step of populating
new rates in the memory store and so it never reaches when the UnknownRate
error can be raised attempting to get the rates from the store the second time.